requirements.txt
# =============================================================================
# Keneya Lens - Dependencies for MedGemma Impact Challenge
# =============================================================================

# Core ML/AI Libraries
# Note: PyTorch 2.1+ required for transformers 4.35+
torch>=2.1.0,<2.5.0
transformers>=4.35.0,<5.0.0
accelerate>=0.24.0
sentence-transformers>=2.2.2

# HAI-DEF Model Support
bitsandbytes>=0.41.0  # For 8-bit/4-bit quantization
peft>=0.7.0  # For LoRA fine-tuning
safetensors>=0.4.0  # Efficient model loading

# Vector Database
chromadb>=0.4.15

# PDF Processing
pypdf2>=3.0.1
pdfplumber>=0.10.3

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0,<2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.1.0
huggingface_hub>=0.19.0

# Image Processing (for medical image uploads)
Pillow>=10.1.0

# System Monitoring (for benchmarking)
psutil>=5.9.0

# Additional utilities
tqdm>=4.66.0

# Optional: Medical embeddings (uncomment if needed)
# pubmedbert>=0.1.0
